refactor(filtering): centralize and streamline electrode filtering
- Removed unused filter function from `blech_process_utils.py`.
- Relocated `get_filtered_electrode` function from `clustering.py` to `read_file.py` for better organization.
- Commented out redundant filter call in `blech_process_utils.py`.
- Updated `read_file.py` to include filtering logic directly within its data processing.
